requirements.txt
# Django
Django==4.2.7

# Database
mysqlclient==2.2.0

# Authentication & Security
PyJWT==2.8.0

# Forms & Validation
django-crispy-forms==2.1
crispy-bootstrap5==2.0.0

# API (if needed for future expansion)
djangorestframework==3.14.0

# Date & Time utilities
python-dateutil==2.8.2
pytz==2023.3

# Environment variables
python-decouple==3.8

# Development tools
django-debug-toolbar==4.2.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.11.0

# Excel/CSV handling (for reports)
openpyxl==3.1.2
pandas==2.1.3